line 151, 152
Hello guys, I'm trying to use this library and I found some weird issue which is not showing a text inside of circle.
Somehow canvas.getWidth and canvas.getHeight returns whole screen width and height instead of widget's width and height so text didn't show. I tried to change it to use layout's width and height and it shows.